backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.3
pydantic==2.5.3
python-multipart==0.0.6
Pillow==10.2.0
scikit-learn==1.4.0
scipy==1.12.0
python-dotenv==1.0.0

# MongoDB
motor==3.3.2
pymongo==4.6.1
